added rmdir

+++ OS-MiniProject/src/rmdir.h
@@ -5,50 +5,47 @@
 #include<errno.h>
 #include<sys/types.h>
 #include<dirent.h>
-#define RESET	"\x1b[0m"
-#define RED		"\x1b[31m"
-#define BLUE	"\x1b[34m"
-
-int Isinteractive(char *dir,char *pathname);
+//prototype of the function needs to be defined above 
+int Isinteractive(char *dir);
 void nonemptydirectoryrecursively(char *dir);
 int isfile(char *filename);
+int emptydirectory(char *dir);
+
 
 
 //argc-counts the no of lines till the input is taken
 //argv-stores the input in form of an array
 
- // function to be called in main
-int is_interactive(char *dir){
-   //creates a directory by using inbuilt opendir function
-	DIR *directory = opendir(dir);
-    
-    //if directory is not empty ,it confirms again to delete it or not 
-     if(directory!=NULL){
-        printf(BLUE"Are u sure u want to delete the non empty directory(y/n)?"RESET);
-       
-		char a;
-		scanf("%c",&a);
-       
-        if(a == 'y'  || a == 'Y'){
-           
-			nonemptydirectoryrecursively(dir);
-                
-        }
-       
-       else{
-       
+
+int emptydirectory(char *dir){
+     //opens the directory
+		DIR *directory = opendir(dir);
+		struct dirent *d;
+		
+		int n=0;
+    //checking whether the directory is empty or non empty
+	while(d=readdir(directory)!=NULL){
+		
+		if(++n>2)
+			break;
+    }
+	//if directory is null
+	if(directory == NULL){
+			printf("directory does not exists\n");
 			exit(0);
-       
-       }
     }
-       
-    //if empty directory then it deletes simply by using rmdir function
+    //if directory is empty
+    else if(n<=2){
+			rmdir(dir);
+    }
+    //if directory is non empty,then gives error
     else{
-        rmdir(dir);
-    }
-       
+			printf("error deleting directory\n");
+            exit(0);
+
+	}
 	return 0;
-      
+        
 }
        
 // function for deleting a non empty directory 
@@ -61,12 +58,8 @@ void nonemptydirectoryrecursively(char *dir){
 	//using the scandir function
 	count = scandir(dir, &entry, NULL, alphasort); 
 	//changes the path of directory after every recursive call         
-	int n = chdir(path);
-	if( n != 0 ){
-		//If directory doesn't exist
-		printf(RED"bash: cd: cn: No such file or directory\n"RESET);
-		return;
- 	}
+	chdir(path);
+	// error handling required
 	for(int i=0;i<count;i++){
 	   //checks whether the given entry is a file or a directory
 		if(isfile(entry[i]->d_name) == 0){
@@ -92,7 +85,7 @@ int isfile(char *filename){
     // close the directory stream
      closedir(directory);
       return 0;
-    }
+    }//if the given entry is a file
 	else if(errno == ENOTDIR){
 
       return 1;
@@ -102,4 +95,51 @@ int isfile(char *filename){
 	else {
       return -1;
   } 
+}
+
+ // function to check whether the given entry is a file or directory
+int Isinteractive(char *dir){
+   //creates a directory by using inbuilt opendir function
+	DIR *directory = opendir(dir);
+	struct dirent *d;
+	int n=0;
+    //checking if directory is empty or not by reading the directory using rmdir function
+	while(d=readdir(directory)!=NULL){
+	  //if n is greater than 2 ,leaves the loop
+		if(++n>2)
+			break;
+    }
+    //empty directory
+    if(n<=2){
+    	
+        rmdir(dir);
+    
+    }
+     //if directory does not exists
+    else if(directory == NULL){
+		printf("directory does not exists\n");
+	
+		exit(0);
+    }
+       
+     else{
+			printf("are u sure u want to delete this non empty directory(y/n)?");
+			char ch;
+			scanf("%c",&ch);
+			
+			//if y , then call the function
+			if(ch == 'y'){
+				nonemptydirectoryrecursively(dir);
+            }
+            //else return
+			else if(ch == 'n'){
+     
+				return 0;
+            } // no other tags included
+			else{
+				printf("invalid character\n");
+            }
+    }
+	return 0;
+      
 }
